Enable and fixes more warnings.

+++ MP42/FFmpegUtils.h
@@ -18,8 +18,8 @@ enum AVCodecID FourCCToCodecID(OSType formatID);
 OSType CodecIDToFourCC(enum AVCodecID codecID);
 
 int remap_layout(AudioChannelLayout *layout, uint64_t in_layout, int count);
-int convert_layout_to_av(AudioChannelLayout *layout, UInt32 layoutSize);
+uint64_t convert_layout_to_av(AudioChannelLayout *layout, UInt32 layoutSize);
 
-int channel_layout_for_channels(AVCodec *codec, int channels);
+uint64_t channel_layout_for_channels(AVCodec *codec, int channels);
 
 #endif /* FFmpegUtils_h */
